BUG: Do not recursively goto()
Since not all steps in Registration workflow have a corresponding secondary tab, _lookupSecondaryStepTabIndex would return incorrect index starting at step 3.

There also is no logic to set the primary tab.

This change adds logic to set the primary tab, and in both cases searches the appropriate QTabBar to find the correct index for the step. If no corresponding tab is found, the index is not changed.

Add a check to goto() so that if the source and destination steps are the same, no action is taken. This avoids recursion when invoking goto() by clicking the tab, or when updating the tab index by invoking goto().