ENH: Update UI state on scene load
HomeLogic - Goto current step. Introduces a `_currentNamesCache` attribute to detect when the parameter node "current" and the real "current" may be out-of-sync and invoke the appropriate `goto` logic.

TableManager, LandmarkLogic - Disconnect existing observers, and add observers to the new nodes.

Observations in Registration and Navigation steps are not supported (yet).