BUG: Fix memory leak on exit ensuring HomeLogic instance is garbage collected
This commit ensures the HomeLogic instance can be deleted by updating the workflow
to use a weak reference to the engine.

It fixes a regression introduced in 5a26bbcd4 (ENH: Introduce Workflow
and Step definition classes).

Since the engine corresponds to the HomeLogic instance itself referenced
in all instances of nested workflows associated with the Patients, Planning,
Registration and Navigation modules, using a weak reference ensure the
logic instance reference count can effectively go to zero and be garbage
collected which in turn ensures the modules can all be garbage collected.

Co-authored-by: David Allemang <david.allemang@kitware.com>